# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "anony-mail API",
    "version": "0.3.0",
    "summary": "REST + Server-Sent Events API for an inbound-only disposable email backend.",
...
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "MIT OR Apache-2.0"
    },
    "contact": {
      "name": "anony-mail source & issues",
      "url": "https://github.com/1mrnewton/anony-mail"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development (default API_BIND_ADDR)"
    },
    {
      "url": "https://{host}",
      "description": "Deployed instance",
      "variables": {
        "host": {
          "default": "api.example.com",
          "description": "Hostname where your anony-mail instance is reachable."
        }
      }
    }
  ],
...
  "paths": {
...
    "/api/addresses/{address}/messages": {
      "parameters": [
        {
          "$ref": "#/components/parameters/AddressPath"
        }
      ],
      "get": {
        "operationId": "listMessages",
        "summary": "List messages in a mailbox",
        "description": "Returns lightweight summaries of the mailbox's messages, newest first (ordered by `received_at` then `id`, both descending). Bodies and attachment metadata are omitted here — fetch a single message for the full content. An existing mailbox with no mail returns `200` with an empty array `[]`; a missing mailbox returns `404`.\n\nPagination: `limit` caps the page size (default 50, max 200). `since` is a keyset cursor — pass the id of the newest message you have seen and only strictly newer messages are returned, which is the cheap way to poll for deltas (e.g. after a push notification). If the `since` message no longer exists (expired, deleted, or dropped by quota), the cursor is ignored and the newest page is returned.",
        "tags": [
          "Messages"
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Maximum number of summaries to return. Values above 200 are clamped.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            }
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "description": "Keyset cursor: return only messages strictly newer than this message id.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Message summaries, newest first (possibly empty).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageSummary"
                  }
                },
                "examples": {
                  "withMail": {
                    "summary": "A mailbox with two messages",
                    "value": [
                      {
                        "id": "6ba7b810-9dad-11d1-80b4-00c04fd430c8",
                        "mail_from": "no-reply@service.test",
                        "subject": "Your verification code",
                        "received_at": "2026-07-01T11:40:12.500000Z",
                        "has_attachments": false,
                        "seen": false
                      },
                      {
                        "id": "550e8400-e29b-41d4-a716-446655440000",
                        "mail_from": "billing@shop.test",
                        "subject": "Invoice #1042",
                        "received_at": "2026-07-01T11:24:10.000000Z",
                        "has_attachments": true,
                        "seen": true
                      }
                    ]
                  },
                  "empty": {
                    "summary": "Existing mailbox, no mail yet",
                    "value": []
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/MailboxNotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "delete": {
        "operationId": "clearInbox",
        "summary": "Delete all messages in a mailbox",
        "description": "Deletes every message (and their attachments) in the mailbox in one call. The mailbox itself, its expiry, and its owner token are untouched. Requires the owner token.",
        "tags": [
          "Messages"
        ],
        "security": [
          {
            "ownerToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Inbox cleared.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClearInboxResponse"
                },
                "example": {
                  "deleted": 4
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/MailboxNotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/addresses/{address}/messages/{id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/AddressPath"
        },
        {
          "$ref": "#/components/parameters/MessageIdPath"
        }
      ],
      "get": {
        "operationId": "getMessage",
        "summary": "Get a full message",
        "description": "Returns a single message including its text and/or HTML bodies and attachment metadata. Attachment bytes are not inlined — download them individually via the attachments endpoint using each attachment's `id`.\n\nSecurity note: `html_body` is untrusted, attacker-controlled markup (anyone can email the inbox). Render it only in a sandboxed iframe or after sanitization — see the schema description.",
        "tags": [
          "Messages"
        ],
        "responses": {
          "200": {
            "description": "The full message.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoredMessage"
                },
                "examples": {
                  "withAttachment": {
                    "summary": "Message with one PDF attachment",
                    "value": {
                      "id": "550e8400-e29b-41d4-a716-446655440000",
                      "mailbox_address": "a1b2c3d4e5@example.com",
                      "mail_from": "billing@shop.test",
                      "subject": "Invoice #1042",
                      "message_date": "2026-07-01T11:24:05.000000Z",
                      "text_body": "Thanks for your order. Your invoice is attached.",
                      "html_body": "<p>Thanks for your order. Your invoice is attached.</p>",
                      "raw_size": 20841,
                      "received_at": "2026-07-01T11:24:10.000000Z",
                      "seen": false,
                      "attachments": [
                        {
                          "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                          "filename": "invoice-1042.pdf",
                          "content_type": "application/pdf",
                          "size": 18240
                        }
                      ]
                    }
                  },
                  "textOnly": {
                    "summary": "Plain-text message, no attachments",
                    "value": {
                      "id": "6ba7b810-9dad-11d1-80b4-00c04fd430c8",
                      "mailbox_address": "a1b2c3d4e5@example.com",
                      "mail_from": "no-reply@service.test",
                      "subject": "Your verification code",
                      "message_date": "2026-07-01T11:40:08.000000Z",
                      "text_body": "Your verification code is 123456. It expires in 10 minutes.",
                      "html_body": null,
                      "raw_size": 512,
                      "received_at": "2026-07-01T11:40:12.500000Z",
                      "seen": false,
                      "attachments": []
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/MessageNotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "delete": {
        "operationId": "deleteMessage",
        "summary": "Delete a single message",
        "description": "Deletes one message (and its attachments) from the mailbox. Requires the owner token. Returns `204 No Content` on success.",
        "tags": [
          "Messages"
        ],
        "security": [
          {
            "ownerToken": []
          }
        ],
        "responses": {
          "204": {
            "description": "Message deleted."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/MessageNotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/addresses/{address}/messages/{id}/read": {
      "parameters": [
        {
          "$ref": "#/components/parameters/AddressPath"
        },
        {
          "$ref": "#/components/parameters/MessageIdPath"
        }
      ],
      "post": {
        "operationId": "markMessageRead",
        "summary": "Mark a message as read",
        "description": "Sets the message's `seen` flag. Idempotent — marking an already-read message succeeds. Open like the read endpoints (knowing the address already grants read access), so no owner token is required.",
        "tags": [
          "Messages"
        ],
        "responses": {
          "204": {
            "description": "Message marked read."
          },
          "404": {
            "$ref": "#/components/responses/MessageNotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/addresses/{address}/messages/{id}/raw": {
      "parameters": [
        {
          "$ref": "#/components/parameters/AddressPath"
        },
        {
          "$ref": "#/components/parameters/MessageIdPath"
        }
      ],
      "get": {
        "operationId": "getRawMessage",
        "summary": "Download the original message (.eml)",
        "description": "Returns the message's original RFC 5322 bytes as received over SMTP (dot-unstuffed), served as `message/rfc822` with an attachment disposition (`{id}.eml`) and `X-Content-Type-Options: nosniff`. Only available when the server runs with `STORE_RAW_MESSAGE=true`; otherwise — and for messages delivered before the toggle was enabled — this endpoint returns `404`.",
        "tags": [
          "Messages"
        ],
        "responses": {
          "200": {
            "description": "The original message bytes.",
            "headers": {
              "Content-Disposition": {
                "description": "Attachment disposition named after the message id.",
                "schema": {
                  "type": "string"
                },
                "example": "attachment; filename=\"550e8400-e29b-41d4-a716-446655440000.eml\""
              },
              "X-Content-Type-Options": {
                "description": "Always `nosniff`.",
                "schema": {
                  "type": "string",
                  "enum": [
                    "nosniff"
                  ]
                }
              }
            },
            "content": {
              "message/rfc822": {
                "schema": {
                  "type": "string",
                  "format": "binary",
                  "description": "Raw RFC 5322 message bytes."
                }
              }
            }
          },
          "404": {
            "description": "Message not found, or raw retention is disabled.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "raw message not available (missing, or raw retention is disabled)"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/addresses/{address}/messages/{id}/attachments/{attachment_id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/AddressPath"
        },
        {
          "$ref": "#/components/parameters/MessageIdPath"
        },
        {
          "$ref": "#/components/parameters/AttachmentIdPath"
        }
      ],
      "get": {
        "operationId": "getAttachment",
        "summary": "Download an attachment",
        "description": "Streams the raw bytes of an attachment. The response's `Content-Type` is the attachment's stored MIME type, except that browser-active types (`text/html`, `application/xhtml+xml`, `image/svg+xml`, `text/xml`, `application/xml`) are served as `application/octet-stream` so a crafted attachment cannot execute on the API origin — the real type remains available in the message's attachment metadata. Every response carries `X-Content-Type-Options: nosniff`. `Content-Disposition` is set to `attachment` with a sanitized ASCII filename, plus an RFC 5987 `filename*=UTF-8''…` form when the original name is non-ASCII.",
        "tags": [
          "Messages"
        ],
        "responses": {
          "200": {
            "description": "Raw attachment bytes.",
            "headers": {
              "Content-Type": {
                "description": "The attachment's stored MIME type, e.g. `application/pdf`. Browser-active types are replaced with `application/octet-stream`.",
                "schema": {
                  "type": "string"
                }
              },
              "Content-Disposition": {
                "description": "Download disposition with a sanitized ASCII filename, plus an RFC 5987 `filename*` form for non-ASCII names.",
                "schema": {
                  "type": "string"
                },
                "example": "attachment; filename=\"invoice-1042.pdf\""
              },
              "X-Content-Type-Options": {
                "description": "Always `nosniff`: forbids MIME sniffing of the download.",
                "schema": {
                  "type": "string",
                  "enum": [
                    "nosniff"
                  ]
                }
              }
            },
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary",
                  "description": "Raw attachment content."
                }
              }
            }
          },
          "404": {
            "description": "The attachment (or its message/mailbox) was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "attachment not found"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ownerToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Per-mailbox owner token returned once by `POST /api/addresses` (and by rotate). Format: `am_` + 43 base64url characters. Sent as `Authorization: Bearer <owner_token>`."
      },
      "claimToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Per-domain claim token returned once by `POST /api/custom-domains`. Format: `amd_` + 43 base64url characters. Sent as `Authorization: Bearer <claim_token>`."
      },
      "clientToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Client-Token",
        "description": "Signed tier token (12h TTL) from `POST /api/entitlements/verify` (unattested) or `POST /api/client/attest` / `assert` (attested, App Attest). Optional: on enforcing servers a request without it is treated as free tier — unless the server reports `attestation.required`, in which case mutations demand an attested one. `Authorization` stays reserved for owner/claim tokens, so both can be sent together."
      }
    },
    "parameters": {
      "AddressPath": {
        "name": "address",
        "in": "path",
        "required": true,
        "description": "Full email address of the mailbox. Case-insensitive (lowercased server-side). The `@` may be sent literally or percent-encoded as `%40`.",
        "schema": {
          "type": "string",
          "format": "email",
          "examples": [
            "a1b2c3d4e5@example.com"
          ]
        }
      },
      "MessageIdPath": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Message UUID.",
        "schema": {
          "type": "string",
          "format": "uuid",
          "examples": [
            "550e8400-e29b-41d4-a716-446655440000"
          ]
        }
      },
      "AttachmentIdPath": {
        "name": "attachment_id",
        "in": "path",
        "required": true,
        "description": "Attachment UUID (from a message's `attachments[].id`).",
        "schema": {
          "type": "string",
          "format": "uuid",
          "examples": [
            "3fa85f64-5717-4562-b3fc-2c963f66afa6"
          ]
        }
      },
...
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Missing or invalid owner token. The response carries `WWW-Authenticate: Bearer`.",
        "headers": {
          "WWW-Authenticate": {
            "description": "The expected authentication scheme.",
            "schema": {
              "type": "string"
            },
            "example": "Bearer"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "examples": {
              "missing": {
                "summary": "No Authorization header",
                "value": {
                  "error": "missing bearer token"
                }
              },
              "invalid": {
                "summary": "Wrong or rotated-away token",
                "value": {
                  "error": "invalid owner token"
                }
              }
            }
          }
        }
      },
...
      "MailboxNotFound": {
        "description": "No mailbox exists for this address.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "mailbox not found"
            }
          }
        }
      },
      "MessageNotFound": {
        "description": "No such message in this mailbox.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "message not found"
            }
          }
        }
      },
...
      "InternalError": {
        "description": "An unexpected server error occurred. Details are logged server-side.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "internal server error"
            }
          }
        }
      }
    },
    "schemas": {
...
      "ClearInboxResponse": {
        "type": "object",
        "description": "Result of clearing a mailbox.",
        "required": [
          "deleted"
        ],
        "properties": {
          "deleted": {
            "type": "integer",
            "format": "int64",
            "description": "Number of messages that were deleted.",
            "examples": [
              4
            ]
          }
        }
      },
      "MessageSummary": {
        "type": "object",
        "description": "Lightweight message representation used in inbox listings (no bodies or attachment metadata).",
        "required": [
          "id",
          "mail_from",
          "subject",
          "received_at",
          "has_attachments",
          "seen"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Message identifier.",
            "examples": [
              "550e8400-e29b-41d4-a716-446655440000"
            ]
          },
          "mail_from": {
            "type": "string",
            "description": "SMTP envelope sender (MAIL FROM) of the message.",
            "examples": [
              "billing@shop.test"
            ]
          },
          "subject": {
            "type": [
              "string",
              "null"
            ],
            "description": "Message subject, or null if the email had no Subject header.",
            "examples": [
              "Invoice #1042"
            ]
          },
          "received_at": {
            "type": "string",
            "format": "date-time",
            "description": "When the server received and stored the message (UTC).",
            "examples": [
              "2026-07-01T11:24:10.000000Z"
            ]
          },
          "has_attachments": {
            "type": "boolean",
            "description": "True if the message has one or more attachments.",
            "examples": [
              true
            ]
          },
          "seen": {
            "type": "boolean",
            "description": "True once the message was marked read via the read endpoint.",
            "examples": [
              false
            ]
          }
        }
      },
...
      "Error": {
        "type": "object",
        "description": "Uniform error envelope returned on any non-2xx response.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error message.",
            "examples": [
              "mailbox not found"
            ]
          },
          "code": {
            "type": "string",
            "description": "Optional machine-readable rejection code, present on policy errors: `pro_required`, `lifetime_cap`, `client_token_expired`, `client_token_invalid`. More codes may be added; ignore unknown ones.",
            "examples": [
              "pro_required"
            ]
          }
        }
      },
...
    }
  }
}